requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.27.2
python-multipart==0.0.6
openai-whisper==20250625
mysql-connector-python==8.2.0
reportlab==4.0.4
python-dotenv==1.0.0
